requirements.txt
fastapi
uvicorn
numpy
websockets
pydantic
piper-tts
whisperx
